Meredith_Canova/src/load/requirements.txt
google-cloud-bigquery
google-cloud-pubsub
python-dotenv
pandas